Split calendar view bubbles for branches.
Previously, if a bad PR was failing tests, this would reflect on the status
for the whole site. Now, the main branch has a big bubble, and each other
tested branch has a small bubble. The point is that if some branch/PR has
failing tests, the main bubble will still show green, indicating passing
tests for the main branch on a site.