Fixes `tests` endpoint to return the latests tests that were run within 24 hours in the past from now for a given site matching certain test names rather than looking in the latest run. The point of this is to return a "current" view of the test status, but using the latest run can lead to seemingly missing tests if a run is in progress but has not yet run the test in question.
The endpoint also returned results for test runs arbitrarily back in the
past leading the possibility that sites that have been inactive for long
periods of time would be shown as having current and passing tests.

Furthermore, the method returned results from some arbitrary branch. This
commit fixes it to only return results from main/master.

The format of the returned results is also changes slightly. Previously,
it looked like this:

  {
    "<site_id>": {
        "<test_name>": {
            <test_results_dict>,
            "branch": "<branch>",
            "test_start_time": "<test_start_time>"
        }
     }
  }

whereas now the <test_results_dict> is not mixed with the other keys, but
returned separately:

  {
    "<site_id>": {
        "<test_name>": {
            "results": <test_results_dict>,
            "branch": "<branch>",
            "test_start_time": "<test_start_time>"
        }
     }
  }

This essentially keeps the test dict returned by the db unmodified under
the "<test_name>" key.